fixes for Visual Studio (#10)
* cffCompressor.cc: variable length arrays not supported in MS Visual C++; use new/delete

* cffCompressor.cc: use _setmode to change stdin/stdout from text to binary mode on Windows

On Windows, the standard streams are opened as text mode (e.g. automatic newlines translation, etc). To read/write binary data from stdin/to stdout we need to set the `O_BINARY` flag using `_setmode` function:

https://msdn.microsoft.com/en-us/library/tw4k6df8.aspx

* Makefile.vc: add makefile for Visual Studio's nmake.exe

* compreff.def: add module-definition file containing exported functions for the Windows DLL

* compreffor.__init__: use platform-specific exe and dll names

* cxxCompressor: restore use_lib=False to use executable by default

as Sam wrote in compreffor module's docstring, the executable approach
is currently faster than the shared-lib-via-cytpes method.

We disabled it because the exectuable was crashing on Windows, but
now the latter issue is fixed.