cffCompressor.cc: use `std::thread::hardware_concurrency()` for NUM_THREADS
The current default value is set to 100, which is too large and results in
'oversubscription' (number of threads exceeding available logical cores).
The workload is decomposed into too small chunks, leading to unnecessary
'context switching' of the CPU from one thread to another, that slow the
overall performace.

`hardware_concurrency` function returns the number of hardware threads available
on the system (e.g. CPU cores or hyperthreading units), or 0 if the information
is not available.

http://www.cplusplus.com/reference/thread/thread/hardware_concurrency/