compreffor: export 'compress' function with 'method_python=False' boolean; deprecate Compreffor and Methods classes
We no longer have three methods, but only two (C++ is default and installed by default), hence the `Methods` enum is useless now.

The Compreffor class does not serve any purpose either. The only method is compress; one can only use it once for a single font, and then throw it away.
It's better to just use a 'compress' function.