Only depend on setuptools-git-ls-files for sdist
Some PEP 517 build frontends, like PyPA build, have a feature to check
that build dependencies are present prior to starting the build.

The setuptools-git-ls-files dependency is only needed when constructing
the sdist but not when building a wheel. Some package managers, such as
nixpkgs, only build wheels but still need to bundle and depend on this
to pass validation.

This commit adds some complexity to specify that setuptools-git-ls-files
is only needed when building an sdist by extending setuptools.